Move shared job collector helpers from sidekiq to base class
I am still thinking how to integrate the config without hardcoding a
"name" there, but for now we can share pretty much most methods for
other job metric collectors the same way we share them in the existing
worker adapters implementation.
